Move utility methods from service to own class
- Add more doc comments
