+++ src/Modules/RoundRankingModule/Interfaces/IRoundRankingService.cs
@@ -1,5 +1,4 @@
-﻿using EvoSC.Common.Interfaces.Models;
-using EvoSC.Modules.Official.RoundRankingModule.Models;
+﻿using EvoSC.Modules.Official.RoundRankingModule.Models;
 
 namespace EvoSC.Modules.Official.RoundRankingModule.Interfaces;
 
@@ -61,36 +60,4 @@ public interface IRoundRankingService
     /// </summary>
     /// <returns></returns>
     public Task FetchAndCacheTeamInfoAsync();
-    
-    /// <summary>
-    /// Determines the winning team based on the given checkpoint data.
-    /// </summary>
-    /// <param name="checkpoints"></param>
-    /// <returns></returns>
-    public PlayerTeam GetWinnerTeam(List<CheckpointData> checkpoints);
-
-    /// <summary>
-    /// Determines whether the winner team should be displayed in the widget.
-    /// </summary>
-    /// <param name="checkpoints"></param>
-    /// <returns></returns>
-    public bool HasPlayerInFinish(List<CheckpointData> checkpoints);
-
-    /// <summary>
-    /// Traverses the checkpoint data list and sets the gained points on each entry.
-    /// </summary>
-    /// <param name="checkpoints"></param>
-    public void SetGainedPointsOnResult(List<CheckpointData> checkpoints);
-    
-    /// <summary>
-    /// Traverses the checkpoint data, calculates and sets the time differences to the leading player.
-    /// </summary>
-    /// <param name="checkpoints"></param>
-    public void CalculateAndSetTimeDifferenceOnResult(List<CheckpointData> checkpoints);
-    
-    /// <summary>
-    /// Traverses the checkpoint data and sets the accent color on each entry.
-    /// </summary>
-    /// <param name="checkpoints"></param>
-    public void SetGainedPointsBackgroundColorsOnResult(List<CheckpointData> checkpoints);
 }
